Change the priority for adding coupons when a product is added to 30.
Cart totals are calculated after a product is added, but don't occur until
'woocommerce_add_to_cart:20'. Some plugins, like WooCommerce Subscriptions, use
the cart subtotal to validate a coupon, but since the coupon is added before
the totals are calculated, an error is displayed.

This primarily occurs when using the 'add-to-cart' and 'coupon_code' query
string parameters together.